refactor: :recycle: divide shared types to separate files
