+++ src/shared/types/Document/Attribute.ts
@@ -1,0 +1,5 @@
+export type Attribute = {
+  id: number;
+  name: string;
+  required: boolean;
+};
