+++ src/shared/types/Document/DocumentType.ts
@@ -1,0 +1,8 @@
+import { Attribute } from './Attribute';
+
+// document-types / attributes
+export type DocumentType = {
+  id: number;
+  name: string;
+  attributes: Attribute[];
+};
