+++ src/shared/types/Document/DocumentVersion.ts
@@ -1,0 +1,13 @@
+import { Signature } from '../Signature';
+import { Value } from './Value';
+
+export type DocumentVersion = {
+  id: number;
+  versionId: number;
+  title: string;
+  description: string;
+  createdAt: Date;
+  values: Value[];
+  base64Content: string;
+  signatures: Signature[];
+};
