+++ src/shared/types/Document/Value.ts
@@ -1,0 +1,4 @@
+export type Value = {
+  attributeName: string;
+  value: string;
+};
