+++ src/shared/types/Signature.ts
@@ -1,0 +1,10 @@
+// signature-controller
+
+import { User } from './User/User';
+
+// POST /sign/{id}
+export type Signature = {
+  hash: number;
+  placeholderTitle: string;
+  user: User;
+};
