+++ src/shared/types/User/Role.ts
@@ -1,0 +1,4 @@
+export type Role = {
+  id: number;
+  name: string;
+};
