+++ src/shared/types/User/User.ts
@@ -1,0 +1,10 @@
+import { Role } from './Role';
+
+// user-controller
+export type User = {
+  id: number;
+  name: string;
+  surname: string;
+  email: string;
+  roles: Role[];
+};
