+++ src/shared/types/index.ts
@@ -1,0 +1,12 @@
+export type { Document } from './Document/Document';
+export type { DocumentType } from './Document/DocumentType';
+export type { DocumentVersion } from './Document/DocumentVersion';
+
+export type { User } from './User/User';
+export type { Role } from './User/Role';
+
+export type { Attribute } from './Document/Attribute';
+export type { Sign } from './Sign';
+export type { SignVoting } from './SignVoting';
+export type { Signature } from './Signature';
+export type { Value } from './Document/Value';
