+++ src/stores/UserStore.ts
@@ -1,4 +1,4 @@
-import { User } from '@/shared/types/sharedTypes';
+import { User } from '@/shared/types';
 import { makeAutoObservable } from 'mobx';
 class UserStore {
   users: User[] = [];
